Move network logic to dedicated hook
