+++ src/@next/components/molecules/CachedImage/types.ts
@@ -2,5 +2,5 @@ export interface IProps {
   url: string;
   url2x?: string;
   alt?: string;
-  children?: React.ReactNode;
+  children?: React.ReactElement;
 }
